Player Character
Created a basic humanoid character in MakeHuman, rigged it and placed into the project.

Used a free Mocap Pack from the Unity Asset Store to import animations into the project.

Animations now need to be discussed to see which animations from this pack we will use before proceeding with the rest of the creation of the player character.