ProBuilderBasic and Cut-down Animations
Added ProBuilder Basic from the asset store to aid in level creation.

Created a simple room with a point light to demonstrate the simplicity, and to potentially test game mechanics.

After meeting to discuss animations, a significant number of the animations in the mo-cap library folder have been deleted, with the animations we are likely to use being placed in a temporary "MAYBE" folder.