Lots of small changes and fixes
Fixed camera clipping issue, animation is now masked off so that only the arms move when picking up objects. The player's arms and head rotate with the camera instead of only when the animation plays, so now the arms follow the player as they look around, and there are no more clipping issues with the camera.

Deleted all of the unused motion capture animation files, meaning that the project is now significantly smaller, and the used animation files have been moved to their own folder for easier folder navigation

A number of folder tweaks have been made to clean up the file directory a little.