Stage 2 Fixes for destuctible objects
Objects now search by type when starting the collision and will apply a small responce force to that object. The FPS controller is currently looking for the box collider over the rigedbody
Current issue is that there is issue putting a timer on the collision so that the same object can do a large amount of damage just by bouncing off of the object multiple times in the same second.