Main Menu
Started work on the main menu for the game. Currently has a main menu when the game starts up, then the beginnings of a "play game" screen where the game is set up and started from. Currently all of the game setup controls are functional

The only thing left to do with this menu is to allow players to press A on a gamepad to join the game. This will add players to a list, which will then determine how many instances of player characters are spawned when the game is loaded.

As well as this menu; options, pause, and credits menus will also be created.