Controls Overhaul and Rewrite & Game Timer
Worked together with Jake to do a complete re-write of the control scripts to make use of new input manager inputs.

Now all controllers should work for all four players, however since multiplayer has been implemented there is now a bug with the object pickup, and the script may require a complete refactor in order to support multiple players.

Implemented a Game Timer set with the drop-down box in the game setup screen. When the scene loads, the timer starts, and when the timer runs out, an endGame scene loads. This scene will eventually become a scoreboard to show who the winner of the game was.