Object Pickup Fix/Rewrite/Overhaul
Object Pickup is now fixed and works per-player rather than per-object after a large overhaul of the code. Each player in the game now should be able to pick up objects, instead of just player one. The object pickup and throw also now works with the controller inputs. X to pickup, B to drop, and RT to throw. Throwforce and Player Strength-factor still need to be added to the new pickup and throw code.

This means that to make an object interactible, give it a mass, and tag it with PhysicsObject. There is no longer need to add ThrowObjectG script as a component

In addition, a basic "googly eye" has been created and saved as a prefab, it is currently buggy and the pupil having a rigidbody messes with the jump of the player, so it has been left out of any scenes for now.