Large Menu Additions ATTEMPT 2
***Re-committed in an attempt to resolve conflicts between unity scenes***

Added "Back" functionality to all menus, so B button on controller or Esc on keyboard can be used to return to previous menus, or to remove a player.

Added a context switch to move between player setup and game setup to fix a bug where the player could get stuck on the game setup UI.

Added a rudimentary pause menu with Resume and Quit currently functional.

Added an Exit Game button to the main menu ready for when the game is built as an executable, so the player can exit the game when they want to.