Player Sound Effects Implemented and lots of fixes/improvements
There are so many little changes in this commit that I have a feeling this is not going to merge nicely, and the resulting pull request may have to be closed.

This version implements player sounds for footsteps, throwing, dropping, pickup, jump, and land.

Problems with the multiple players has been solved as it appears an issue occurred with merging this fix last time, however it should all be done now.

Recreated the player prefabs as they had not been applied in a long time.

Took out the game music for now until suitable music has been found.